# Speed up CeleryExecutor when sending only a single task to Celery