@ashb 2020-05-13
(Somewhat surprised to find creating a process takes 100ms. It should be a fork which should)

Change makes sense though