@auvipy 2020-05-14
in the next release of celery 4.4.3 you can optionally use cythonize version of py-amqp in some part which should improve some performence.